更新
1.修改凸透镜错误
2.技能释放回复能量支持float

+++ Client/Assets/Scripts/Battle/Skill.cs
@@ -27,7 +27,7 @@ public class Skill : MonoBehaviour
     
     public bool targetSelf;//目标是自己还是对方
 
-    public int manaProduce;//魔力生成
+    public float manaProduce;//魔力生成
     
     public bool ifActive;//是否为主动技能
     public int buffID;
